Move database initialization into SQLZenStore
