+++ src/zenml/cli/server.py
@@ -191,7 +191,7 @@ def up(
     assert gc.store is not None
 
     if not blocking:
-        from zenml.zen_stores.base_zen_store import (
+        from zenml.constants import (
             DEFAULT_PASSWORD,
             DEFAULT_USERNAME,
         )
