+++ src/zenml/zen_stores/base_zen_store.py
@@ -12,8 +12,7 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 """Base Zen Store implementation."""
-import os
-from abc import ABC, abstractmethod
+from abc import ABC
 from typing import (
     Any,
     Callable,
@@ -32,20 +31,14 @@ from pydantic import BaseModel
 from requests import ConnectionError
 
 import zenml
-from zenml.analytics.utils import analytics_disabler
 from zenml.config.global_config import GlobalConfiguration
 from zenml.config.server_config import ServerConfiguration
 from zenml.config.store_config import StoreConfiguration
 from zenml.constants import (
-    ENV_ZENML_DEFAULT_USER_NAME,
-    ENV_ZENML_DEFAULT_USER_PASSWORD,
-    ENV_ZENML_DEFAULT_WORKSPACE_NAME,
     IS_DEBUG_ENV,
 )
 from zenml.enums import (
-    AuthScheme,
     SecretsStoreType,
-    StackComponentType,
     StoreType,
 )
 from zenml.exceptions import AuthorizationException
@@ -53,15 +46,11 @@ from zenml.logger import get_logger
 from zenml.models import (
     ServerDatabaseType,
     ServerModel,
-    StackFilter,
     StackResponse,
     UserFilter,
-    UserRequest,
     UserResponse,
     WorkspaceResponse,
 )
-from zenml.models.v2.core.component import InternalComponentRequest
-from zenml.models.v2.core.stack import InternalStackRequest
 from zenml.utils.proxy_utils import make_proxy_class
 from zenml.zen_stores.enums import StoreEvent
 from zenml.zen_stores.secrets_stores.base_secrets_store import BaseSecretsStore
@@ -74,11 +63,6 @@ from zenml.zen_stores.secrets_stores.sql_secrets_store import (
 from zenml.zen_stores.zen_store_interface import ZenStoreInterface
 
 logger = get_logger(__name__)
-
-DEFAULT_USERNAME = "default"
-DEFAULT_PASSWORD = ""
-DEFAULT_WORKSPACE_NAME = "default"
-DEFAULT_STACK_AND_COMPONENT_NAME = "default"
 
 
 @make_proxy_class(SecretsStoreInterface, "_secrets_store")
@@ -299,19 +283,6 @@ class BaseZenStore(
 
     def _initialize_database(self) -> None:
         """Initialize the database on first use."""
-        default_workspace = self._get_or_create_default_workspace()
-
-        config = ServerConfiguration.get_server_config()
-        # If the auth scheme is external, don't create the default user
-        if config.auth_scheme != AuthScheme.EXTERNAL:
-            try:
-                _ = self._default_user
-            except KeyError:
-                self._create_default_user()
-
-        self._get_or_create_default_stack(
-            workspace=default_workspace,
-        )
 
     @property
     def url(self) -> str:
@@ -452,24 +423,6 @@ class BaseZenStore(
         """
         return self.get_store_info().is_local()
 
-    def _get_or_create_default_stack(
-        self, workspace: WorkspaceResponse
-    ) -> StackResponse:
-        try:
-            return self._get_default_stack(
-                workspace_id=workspace.id,
-            )
-        except KeyError:
-            return self._create_default_stack(
-                workspace_id=workspace.id,
-            )
-
-    def _get_or_create_default_workspace(self) -> WorkspaceResponse:
-        try:
-            return self._default_workspace
-        except KeyError:
-            return self._create_default_workspace()
-
     # --------------
     # Event Handlers
     # --------------
@@ -506,152 +459,6 @@ class BaseZenStore(
                     exc_info=True,
                 )
 
-    # ------
-    # Stacks
-    # ------
-
-    def _create_default_stack(
-        self,
-        workspace_id: UUID,
-    ) -> StackResponse:
-        """Create the default stack components and stack.
-
-        The default stack contains a local orchestrator and a local artifact
-        store.
-
-        Args:
-            workspace_id: ID of the workspace to which the stack
-                belongs.
-
-        Returns:
-            The model of the created default stack.
-        """
-        with analytics_disabler():
-            workspace = self.get_workspace(workspace_name_or_id=workspace_id)
-
-            logger.info(
-                f"Creating default stack in workspace {workspace.name}..."
-            )
-
-            orchestrator = self.create_stack_component(
-                component=InternalComponentRequest(
-                    # Passing `None` for the user here means the orchestrator
-                    # is owned by the server, which for RBAC indicates that
-                    # everyone can read it
-                    user=None,
-                    workspace=workspace.id,
-                    name=DEFAULT_STACK_AND_COMPONENT_NAME,
-                    type=StackComponentType.ORCHESTRATOR,
-                    flavor="local",
-                    configuration={},
-                ),
-            )
-
-            artifact_store = self.create_stack_component(
-                component=InternalComponentRequest(
-                    # Passing `None` for the user here means the stack is owned
-                    # by the server, which for RBAC indicates that everyone can
-                    # read it
-                    user=None,
-                    workspace=workspace.id,
-                    name=DEFAULT_STACK_AND_COMPONENT_NAME,
-                    type=StackComponentType.ARTIFACT_STORE,
-                    flavor="local",
-                    configuration={},
-                ),
-            )
-
-            components = {
-                c.type: [c.id] for c in [orchestrator, artifact_store]
-            }
-
-            stack = InternalStackRequest(
-                # Passing `None` for the user here means the stack is owned by
-                # the server, which for RBAC indicates that everyone can read it
-                user=None,
-                name=DEFAULT_STACK_AND_COMPONENT_NAME,
-                components=components,
-                workspace=workspace.id,
-            )
-            return self.create_stack(stack=stack)
-
-    def _get_default_stack(
-        self,
-        workspace_id: UUID,
-    ) -> StackResponse:
-        """Get the default stack for a user in a workspace.
-
-        Args:
-            workspace_id: ID of the workspace.
-
-        Returns:
-            The default stack in the workspace.
-
-        Raises:
-            KeyError: if the workspace or default stack doesn't exist.
-        """
-        default_stacks = self.list_stacks(
-            StackFilter(
-                workspace_id=workspace_id,
-                name=DEFAULT_STACK_AND_COMPONENT_NAME,
-            )
-        )
-        if default_stacks.total == 0:
-            raise KeyError(
-                f"No default stack found in workspace {workspace_id}."
-            )
-        return default_stacks.items[0]
-
-    # -----
-    # Users
-    # -----
-
-    @property
-    def _default_user_name(self) -> str:
-        """Get the default user name.
-
-        Returns:
-            The default user name.
-        """
-        return os.getenv(ENV_ZENML_DEFAULT_USER_NAME, DEFAULT_USERNAME)
-
-    @property
-    def _default_user(self) -> UserResponse:
-        """Get the default user.
-
-        Returns:
-            The default user.
-
-        Raises:
-            KeyError: If the default user doesn't exist.
-        """
-        user_name = self._default_user_name
-        try:
-            return self.get_user(user_name)
-        except KeyError:
-            raise KeyError(f"The default user '{user_name}' is not configured")
-
-    def _create_default_user(self) -> UserResponse:
-        """Creates a default user.
-
-        Returns:
-            The default user.
-        """
-        user_name = os.getenv(ENV_ZENML_DEFAULT_USER_NAME, DEFAULT_USERNAME)
-        user_password = os.getenv(
-            ENV_ZENML_DEFAULT_USER_PASSWORD, DEFAULT_PASSWORD
-        )
-
-        logger.info(f"Creating default user '{user_name}' ...")
-        new_user = self.create_user(
-            UserRequest(
-                name=user_name,
-                active=True,
-                password=user_password,
-            )
-        )
-        return new_user
-
     def get_external_user(self, user_id: UUID) -> UserResponse:
         """Get a user by external ID.
 
@@ -668,47 +475,6 @@ class BaseZenStore(
         if users.total == 0:
             raise KeyError(f"User with external ID '{user_id}' not found.")
         return users.items[0]
-
-    # --------
-    # Workspaces
-    # --------
-
-    @property
-    def _default_workspace_name(self) -> str:
-        """Get the default workspace name.
-
-        Returns:
-            The default workspace name.
-        """
-        return os.getenv(
-            ENV_ZENML_DEFAULT_WORKSPACE_NAME, DEFAULT_WORKSPACE_NAME
-        )
-
-    @property
-    def _default_workspace(self) -> WorkspaceResponse:
-        """Get the default workspace.
-
-        Returns:
-            The default workspace.
-
-        Raises:
-            KeyError: if the default workspace doesn't exist.
-        """
-        workspace_name = self._default_workspace_name
-        try:
-            return self.get_workspace(workspace_name)
-        except KeyError:
-            raise KeyError(
-                f"The default workspace '{workspace_name}' is not configured"
-            )
-
-    @abstractmethod
-    def _create_default_workspace(self) -> WorkspaceResponse:
-        """Creates a default workspace.
-
-        Returns:
-            The default workspace.
-        """
 
     class Config:
         """Pydantic configuration class."""
