+++ tests/harness/deployment/server_docker_compose.py
@@ -264,7 +264,7 @@ services:
         Raises:
             RuntimeError: If the deployment is not running.
         """
-        from zenml.zen_stores.base_zen_store import (
+        from zenml.constants import (
             DEFAULT_PASSWORD,
             DEFAULT_USERNAME,
         )
