+++ tests/harness/deployment/server_local.py
@@ -162,7 +162,7 @@ class ServerLocalTestDeployment(BaseTestDeployment):
         Raises:
             RuntimeError: If the deployment is not running.
         """
-        from zenml.zen_stores.base_zen_store import (
+        from zenml.constants import (
             DEFAULT_PASSWORD,
             DEFAULT_USERNAME,
         )
